#528: PR Fixes
