+++ src/main/scala/za/co/absa/hyperdrive/trigger/api/rest/auth/LdapGrantedAuthoritiesMapper.scala
